archive/piper-morgan-0.1.1/requirements.txt
# Core dependencies
langchain
langchain-openai
langchain-community
openai
anthropic
chromadb
python-dotenv

# GitHub integration
PyGithub

# Document processing
pypdf2
python-docx

# Web interface
streamlit

# HTTP client (for anthropic)
# httpx==0.28.1

# Avoid version conflicts
urllib3==1.26.18

# Sentence Transformers for embeddings
sentence-transformers
